batch: Move realized entry storage
The batch merge module applies merge constraints, enumerates candidates,
and stores realized line provenance for materialized content.

The realized-entry storage code is a compact buffer abstraction with
callers in source advancement and storage, so keeping its public names on
merge makes the merge module carry a second responsibility.

This commit addresses that by moving realized entries and chunk streaming
into batch.realized_entries, updating production callers, and keeping merge
imports private.

Subsequent commits will update direct realized-entry tests and guard the
new boundary.
