+++ src/git_stage_batch/batch/source_advancement.py
@@ -13,10 +13,8 @@ from ..utils.repository_buffers import (
 )
 from ..utils.git import get_git_repository_root_path
 from .lineage import BatchSourceLineage, LineageRun
-from .merge import (
-    apply_presence_constraints,
-    realized_entry_content_chunks,
-)
+from .merge import apply_presence_constraints
+from .realized_entries import realized_entry_content_chunks
 from .ownership import BatchOwnership, remap_batch_ownership_with_lineage
