+++ src/git_stage_batch/batch/storage.py
@@ -51,7 +51,8 @@ from ..utils.git import (
 )
 from ..utils.paths import get_batch_metadata_file_path
 from ..utils.text import normalize_line_sequence_endings
-from .merge import realized_entry_content_chunks, satisfy_constraints
+from .merge import satisfy_constraints
+from .realized_entries import realized_entry_content_chunks
 
 if TYPE_CHECKING:
     from .ownership import BatchOwnership, AbsenceClaim
